requirements.txt
amqp==2.2.2
billiard==3.5.0.3
celery==4.1.0
certifi==2018.1.18
chardet==3.0.4
dj-database-url==0.5.0
Django==2.0.4
django-filter==1.1.0
djangorestframework==3.8.2
flake8==3.5.0
gunicorn==19.7.1
idna==2.6
kombu==4.1.0
Markdown==2.6.11
mccabe==0.6.1
praw==5.4.0
prawcore==0.14.0
psycopg2==2.7.4
pycodestyle==2.3.1
pyflakes==1.6.0
python-decouple==3.1
pytz==2018.4
redis==2.10.6
requests==2.18.4
update-checker==0.16
urllib3==1.22
vine==1.1.4
whitenoise==3.3.1
